move test helpers to conftest.py
Extract create_book() and create_booklog_csv_row() from test_core.py
into tests/conftest.py so they can be shared across test files.
