+++ tests/test_core.py
@@ -1,6 +1,6 @@
+from conftest import create_book, create_booklog_csv_row
+
 from booklog_sync.core import (
-    Book,
-    BooklogCSVRow,
     _sanitize_filename,
     build_id_book_index,
     convert_csv,
@@ -8,42 +8,6 @@ from booklog_sync.core import (
     generate_filename,
     save_book_to_markdown,
 )
-
-
-def create_booklog_csv_row(props: dict[str, str] | None = None) -> BooklogCSVRow:
-    if props is None:
-        props = {}
-
-    default_csv_row: BooklogCSVRow = {
-        "item_id": "1000000000",
-        "title": "テストタイトル",
-        "author": "テスト作者名",
-        "isbn13": "9784000000001",
-        "publisher": "テスト出版社",
-        "publish_year": "2020",
-        "status": "読み終わった",
-        "rating": "5",
-    }
-
-    return {**default_csv_row, **props}
-
-
-def create_book(props: dict[str, str] | None = None) -> Book:
-    if props is None:
-        props = {}
-
-    default_book: Book = {
-        "item_id": "1000000000",
-        "title": "テストタイトル",
-        "author": "テスト作者名",
-        "isbn13": "9784000000001",
-        "publisher": "テスト出版社",
-        "publish_year": "2020",
-        "status": "読み終わった",
-        "rating": 5,
-    }
-
-    return {**default_book, **props}
 
 
 def test_convert_row_to_properties():
